// src/app/Model/User.ts
export class User {
    name : string;
    phone: string;
    title: string;
    email: string;
    password: string;
    mobile:string;
    contactId:string;
    producerId:string;
    prefix:string;
    fullAddress2: string;
    address1: string;
    address2: string;
    city: string;
    state: string;
    postalCode: string;
    country: string;
    fullAddress1: string;
    mailAddress1: string;
    mailAddress2: string;
    mailCity: string;
    mailState: string;
    mailPostalCode: string;
    mailCountry: string;
    mailFullAddress1: string;
  }
  
  